// src/api/generated/model/carResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * MiniCRM API
 * API Documentation for MiniCRM Application
 * OpenAPI spec version: 1.0.0
 */
import type { CarResponseStatus } from './carResponseStatus';

export interface CarResponse {
  id?: number;
  licensePlate?: string;
  model?: string;
  brand?: string;
  horsepower?: string;
  status?: CarResponseStatus;
  createdAt?: string;
  updatedAt?: string;
}